bugfix(web-ui): fix photo upload on android devices
