+++ services/web-ui/src/apis.ts
@@ -1,5 +1,6 @@
-import { Configuration, ProfileApi } from "src/api";
+import { Configuration, PhotosApi, ProfileApi } from "src/api";
 
 const config = new Configuration({ basePath: window.location.origin });
 
 export const profileApi = new ProfileApi(config);
+export const photosApi = new PhotosApi(config);
